Add custom_type constructor and custom types.
custom_types() just wrapps a callable to give it a nice __repr__ for
when type errors are raised.

MinLength, Range, Contains and RegExp illustrates both the creation of
reusable custom type specifications (predicates) and the use of many
arguments to typed_return (i.e. where "," argument separator can be seen
as the conjunction operator).